Feat : 메인 페이지를 로그인 페이지로 변경, AuthProvider, middleware 로직 수정
